pyproject.toml
[tool.poetry]
name = "healthdataintegrationliteraturereview"
version = "0.1.0"
description = ""
authors = ["Marcel Parciak <marcel.parciak@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
bibtexparser = {version = "^2.0.0b3", allow-prereleases = true}


[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
